Break style+explode tests into files for each param type
